[JIT] Enable ModuleDict non-literal indexing
**Summary**
This commit enables indexing into `ModuleDict` using a non-literal
index if the left-hand side of the assignment has an interface type
annotation and if the dictionary contains only modules that implement
that interface, e.g. `module: ModuleInterface = self.module_dict[key]`.

The approach taken by this commit is that the annotation is passed
into `SugaredValue::getitem` as a type hint whose presence is used to
emit a special `prim::ModuleDictIndex` operator and set the output type
of the node appropriately.

**Test Plan**
This commit adds unit tests to `TestModuleContainers` to test this
feature (`test_typed_module_dict`).

ghstack-source-id: e4ce78314e072c48c20d3c6ca0caef2fbcf3842d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45716